refactor: Migrate SentenceTransformerEmbedding class and Semantic Chunker initialization function to models folder
